refactor: feature:main 모듈의 패키지 구조 단순화 (#42)
- `ui` 서브패키지를 제거하고 `MainActivity`, `MainScreen`, `MainViewModel`을 `com.itschristmas.main` 최상위 패키지로 이동하여 구조를 단순화
