+++ feature/main/src/main/java/com/itschristmas/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.itschristmas.main.ui
+package com.itschristmas.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
